+++ src/serialize/xml/normalize.ts
@@ -531,11 +531,11 @@ export class ComponentNormalizer extends BaseXmlNormalizer<Models.Component> {
           children: this._factory.makeForProperty().normalizeIterable(data.properties, options, 'property')
         }
       : undefined
-    const tags: SimpleXml.Element | undefined = spec.supportsComponentTags && data.tags !== undefined && data.tags.length > 0
+    const tags: SimpleXml.Element | undefined = spec.supportsComponentTags && data.tags.size > 0
       ? {
           type: 'element',
           name: 'tags',
-          children: data.tags.map(value => makeTextElement(value, 'tag', normalizedString))
+          children: makeTextElementIter(data.tags, options, 'tag', normalizedString)
         }
       : undefined
     const components: SimpleXml.Element | undefined = data.components.size > 0